Move code away from tiny .go files
If a file has very little Go code and it's all used in a single file,
just move it to that file.

There are more files like these, but these three were the smallest.

Updates #334.
