+++ redis_signal_handle_config.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
+	"syscall"
 	"time"
 
 	"github.com/TykTechnologies/logrus"
@@ -124,4 +125,19 @@ func HandleNewConfiguration(payload string) {
 	}).Info("Initiating configuration reload")
 
 	ReloadConfiguration()
+}
+
+func ReloadConfiguration() {
+	myPID := HostDetails.PID
+	if myPID == 0 {
+		log.Error("No PID found, cannot reload")
+		return
+	}
+
+	log.Info("Sending reload signal to PID: ", myPID)
+
+	callErr := syscall.Kill(myPID, syscall.SIGUSR2)
+	if callErr != nil {
+		log.Error("Process reload failed: ", callErr)
+	}
 }
