--- redis_signal_handle_reload.go
@@ -1,12 +1,0 @@
-package main
-
-import (
-	"github.com/TykTechnologies/logrus"
-)
-
-func HandleReloadMsg() {
-	log.WithFields(logrus.Fields{
-		"prefix": "pub-sub",
-	}).Info("Reloading endpoints")
-	ReloadURLStructure()
-}
